Moved source files to src/ folder. Changed MakeFile.
