+++ Makefile
@@ -1,12 +1,14 @@
 # You can put your build options here
 -include config.mk
 
+IDIR=./src
+
 all: libjsmn.a 
 
 libjsmn.a: jsmn.o
 	$(AR) rc $@ $^
 
-%.o: %.c jsmn.h
+%.o: $(IDIR)/%.c $(IDIR)/jsmn.h
 	$(CC) -c $(CFLAGS) $< -o $@
 
 test: test_default test_strict test_links test_strict_links
